refactor(quota): extract monitor poll CLI owner
