+++ loopx/cli_commands/quota.py
@@ -86,6 +86,7 @@ from .quota_request import (
     quota_detail_sections_from_args,
     validate_quota_command_request,
 )
+from .quota_monitor_poll import record_quota_monitor_poll_for_cli
 from .quota_registration import register_quota_command as register_quota_command
 
 PrintPayload = Callable[
@@ -519,33 +520,6 @@ def _quota_renderer(
     }.get(command, render_quota_markdown)
 
 
-def _receipt_bound_monitor_todo_id(
-    args: argparse.Namespace,
-    *,
-    runtime_root: Path,
-    turn_instance_id: str | None,
-) -> str | None:
-    if not turn_instance_id:
-        return None
-    receipt = find_heartbeat_receipt(
-        runtime_root,
-        goal_id=args.goal_id,
-        agent_id=args.agent_id,
-        turn_instance_id=turn_instance_id,
-    )
-    if receipt is None:
-        raise HeartbeatReceiptIdentityConflictError(
-            "turn-scoped monitor-poll requires a committed heartbeat receipt; "
-            "run quota should-run with the same --turn-instance-id first"
-        )
-    todo_id = heartbeat_receipt_settlement_todo_id(receipt)
-    if todo_id is None:
-        raise HeartbeatReceiptIdentityConflictError(
-            "turn-scoped monitor-poll requires a heartbeat receipt bound to a Todo"
-        )
-    return todo_id
-
-
 def handle_quota_command(
     args: argparse.Namespace,
     *,
@@ -692,41 +666,14 @@ def handle_quota_command(
                         heartbeat_stall_observation = "not_applicable"
                     heartbeat_receipt_ready = True
         elif args.quota_command == "monitor-poll":
-            payload = record_quota_monitor_poll(
-                status_payload,
-                goal_id=args.goal_id,
+            payload = record_quota_monitor_poll_for_cli(
+                args,
+                status_payload=status_payload,
                 registry_path=registry_path,
-                execute=bool(args.execute),
-                source=args.source,
-                reason_summary=args.reason_summary,
-                agent_id=args.agent_id,
-                available_capabilities=args.available_capabilities,
-                todo_id=args.todo_id,
-                target_key=args.target_key,
-                result_hash=args.result_hash,
-                material_change=bool(args.material_change),
-                cadence=args.cadence,
-                next_due_at=args.next_due_at,
-                next_agent_todo=args.next_agent_todo,
-                next_action_kind=args.next_action_kind,
-                next_task_repository=args.next_task_repository,
-                next_required_capabilities=args.next_required_capabilities,
-                next_continuation_policy=args.next_continuation_policy,
-                next_target_key=args.next_target_key,
-                next_user_todo=args.next_user_todo,
-                next_user_task_class=args.next_user_task_class,
-                next_claimed_by=args.next_claimed_by,
+                runtime_root=runtime_root,
                 turn_instance_id=heartbeat_turn_id,
-                receipt_bound_todo_id=_receipt_bound_monitor_todo_id(
-                    args,
-                    runtime_root=runtime_root,
-                    turn_instance_id=heartbeat_turn_id,
-                ),
                 scheduler_execution_context=scheduler_context,
                 operator_inbox_urgency_projector=operator_inbox_urgency_projector,
-                bounded_research_frontier_projector=(
-                    project_live_explore_composition_frontier
-                ),
                 status_reloader=lambda: collect_status(
                     registry_path=registry_path,
                     runtime_root_override=runtime_root_arg,
